Add support for intermediate fractional values.
Solve for real Nerdle problems like "3/7*21=9".

Access the int value through int_value method, which returns a Result.

len() method also now returns a Result since the length of a fraction is not valid (cannot be used in final answer).

Plus and Minus will also fail with a fraction operand (I don't think this can happen in practice because the
equations aren't long enough), but Times and Divide will work, simplifying the result to an integer if possible.